Criada a conexão com MySQL e tabelas criadas com sucesso!

+++ src/main/java/com/meuacervo/meu_acervo/service/EmprestimoService.java
@@ -1,10 +1,12 @@
 package com.meuacervo.meu_acervo.service;
 
+import com.meuacervo.meu_acervo.DTOs.CreateEmprestimoDTO;
 import com.meuacervo.meu_acervo.model.Emprestimo;
 import com.meuacervo.meu_acervo.repository.EmprestimoRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
+import java.time.Instant;
 import java.util.List;
 import java.util.Optional;
 
@@ -13,19 +15,32 @@ public class EmprestimoService {
     @Autowired
     private EmprestimoRepository emprestimoRepository;
 
-    public List<Emprestimo> findAll() {
-        return emprestimoRepository.findAll();
+    public Integer createEmprestimo(CreateEmprestimoDTO createEmprestimoDTO) {
+        var entity = new Emprestimo(
+                createEmprestimoDTO.id(),
+                createEmprestimoDTO.dataEmprestimo(),
+                createEmprestimoDTO.dataDevolucao(),
+                createEmprestimoDTO.livro(),
+                createEmprestimoDTO.colaborador(),
+                Instant.now(),
+                null
+        );
+        var emprestimoSaved = emprestimoRepository.save(entity);
+        return emprestimoSaved.getId();
     }
 
-    public Optional<Emprestimo> findById(Integer id) {
+    public Optional<Emprestimo> findEmprestimoById(Integer id) {
         return emprestimoRepository.findById(id);
     }
 
-    public Emprestimo save(Emprestimo emprestimo) {
-        return emprestimoRepository.save(emprestimo);
+    public List<Emprestimo> listEmprestimos() {
+        return emprestimoRepository.findAll();
     }
 
     public void deleteById(Integer id) {
-        emprestimoRepository.deleteById(id);
+        var emprestimoExists = emprestimoRepository.existsById(id);
+        if (emprestimoExists) {
+            emprestimoRepository.deleteById(id);
+        }
     }
 }
